requirements.txt
boto3==1.40.30
botocore==1.40.30
jmespath==1.0.1
protobuf==6.32.1
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
google-generativeai==0.8.5
psycopg2-binary==2.9.10